[issues]823 add broker_config_request_handler to process broker_config info.
